Adds support for Japanese in the Index
If the data-primary attribute of a data-type=indexterm element, or the
data-primary-sortas attribute if present, starts with a Hiragana or
Katakana character, create the relevant entry in the index and place
this index term there.

This includes normalizing Katakana to Hiragana, removing diacritics, and
normalizing small kana to large kana, as is customary.

If the data-primary starts with a Kanji and data-primary-sortas does
not give the corresponding kana, the term is still placed in the symbols
category, as previously.

This also makes sure that the kana based terms are sorted according to
the customary order, instead of being sorted by Unicode code point.