Adding Title Attribute to <reference> element in content.opf
When testing whether Amazon will ingest our epubs, we've been
receiving errors that state: "Guide title is empty. Item is ignored"
and "This index is empty and has been skipped". I was able to locate the
source of this error in the content.opf file, which is created by opf.xsl.

In content.opf, there is a <guide> element, which contains <reference>
elements as children. There is a <reference> for the Cover, TOC, and
Title Page. The <reference>s for the Cover and TOC each have 3 attributes:
href, type, and title, while the <reference> for the Title Page only has
the first two.

Looking at the logic within opf.xsl, this seems to be because the Title Page
<reference> is set up to be flexible if there is no Title Page within the
book, and so we didn't want to hardcode a title attribute. However, given
that this is preventing ingestion to Amazon as well as the fact that every
O'Reilly title built in Atlas has a title page, I believe it is worth it
to hardcode this attribute.