Removing TOC from Guide
For an unknown reason, the link to the TOC in the <guide> element
is causing an error when Amazon attempts to ingest our EPUBs. Rather
than remove the <guide> element entirely, we decided to remove just
the TOC item to preserve some of the functionality while still allowing
us to send EPUBs to Amazon.