Updating opf.xspec Based on Changes to opf.xsl
We had to change sections of the generate-guide template in opf.xsl
because it was causing errors when submitting EPUBs to Amazon. This
commit updates the opf.xspec test related to generate-guide to match
those changes.

* We are no longer incuding a reference to the TOC, so the test was
changed to not(exists(
* A "title" attribute was required for the reference to the start of
text reference, so an exists( test was added for that

Unfortunately, we have not been able to get this test working, so it
is still marked as pending.