Removing Italics from Titles in XRefs to Appendixes, Chapters, and Parts
To conform with the Chicago Manual of Style, I removed the XSL template
that added an <em> element around the titles generated by xrefs to
Appendixes, Chapters, and Parts. They are now treated like all other
elements in that the title is copied without any surrounding element
(see the template on lines 602-609 of xrefgen.xsl).