Updating XSpec Tests for New TOC Standard
Now that we are including About the Author sections in the TOC, I
updated the XSpec tests to match. To do this I added an additional
test making sure that the abouttheauthor colophon section is included
in the TOC, and adjusted the previous version of the test to indicate
that colophons without an aboutheauthor class will continue to be
excluded.